Do not show prev/next link to posts that are part of the `blog` section
I had to use a nested condition because `and` eveluates all conditions
inside the if statement. This will result in a `nil` pointer dereference
because I want to check the `Section` of `.Prev` or `.Next` but one of
them or both can be `nil` if there are no other/newer/older posts inside
the section.